🌈 style(rename): Rename files
folder: src/config to src/utils
file: config/utils.py to tools.py
