+++ src/client/apfl.py
@@ -4,7 +4,7 @@ from typing import Dict, OrderedDict
 import torch
 
 from fedavg import FedAvgClient
-from src.config.utils import trainable_params
+from src.utils.tools import trainable_params
 
 
 class APFLClient(FedAvgClient):
