+++ src/client/pfedme.py
@@ -6,7 +6,7 @@ import torch
 from torch.optim import Optimizer
 
 from fedavg import FedAvgClient
-from src.config.utils import trainable_params
+from src.utils.tools import trainable_params
 
 
 class pFedMeClient(FedAvgClient):
