+++ src/server/fedap.py
@@ -9,7 +9,7 @@ import numpy as np
 from rich.progress import track
 
 from fedavg import FedAvgServer, get_fedavg_argparser
-from src.config.utils import trainable_params
+from src.utils.tools import trainable_params
 from src.client.fedap import FedAPClient
